Implemented health system
As with Mover, not much needed to be changed relative to planning for this component to be implemented. There was some trickiness with coroutine syntax, but I managed to work everything out in the end. I also added a test for the health component which simply constantly has the player take damage. This is a convenient way to test the damage function, the invincibility period, and the isDead flag all at once, so it's pretty convenient :p