Added weapon ScriptableObject
This class is essentially just a collection of variables, so there isn't
really anything that can go wrong implementing it. It should be noted
however that something I did not account for in planning is that I
needed to add an attribute to the class which allows for a projectile to
be created from the unity asset creation menu. This was simple enough though