Added projectile component
For this commit I was planning to implement the shooter system, but
before that happened I realized that first I needed to create the logic
for the projectiles that the component would actually create. The logic
for explosions has not yet been implemented, but I just needed something
for the Shooter component to fire during testing.