Fully implemented shooter component
Now that I had created the projectile class, I could move on to creating
the shooter class as I had something for it to actually work with. I was originally planning on spawning a new empty gameobject and
attaching the necessary components whenever a projectile was spawned,
but I found that having a pre-defined prefab for any projectiles that
were spawned was a more elegant option. I was also able to fully
implement the logic for the projectiles themselves, including explosion
behaviour and despawning stray projectiles; the only thing left to
implement is damage logic, but that can't really be added until enemy
logic is also implemented.