Implemented player controller
This part is when it all starts coming together!
I've finally added player movement and shooting controls, using unity's
new input system. I had to make some modifications to the shooter class
and the projectile class - making the projectiles spawn with their
position and rotation relative to the player was tricky and something I
didn't account for previously - but it all worked out in the end, and
you can see all the changes I made by clicking on the commit message.