Added and implemented simple enemy controller
As described in the planning document, this simple enemy controller just
locks on to the player (with interpolation) and constantly fires its
weapon. The rate at which the enemy locks on to the player can be
controlled. The next thing to implement is damage, now that something
can damage the player.