Implemented damage... and some other things
This is when I decided to implement damage, since the player actually
has something to be damaged by now. I've also done some work on the
visuals of the game; e.g adding color to the player and enemies,
changing the background color, and zooming out the camera. Next up is
doing some work on cleaning the code, reformatting and commenting
everything which might not be written very well.