Added walls
Before I clean everything up, I just wanted to add something that I
couldn't find any other time to do, which was adding collision walls
around the arena. This code was "inspired" by a Unity forum post, and
it adapts to the screen's resolution as it goes around the area which
the camera covers.