Cleaned up and commented everything
I've cleaned up and commented everything which wasn't commented. I've
also changed the enemy behaviour so that the turn speed is independent
of framerate. Something I also forgot to mention in the previous commit
is that I implemented dynamic colors for the player and enemies, meaning
that they turn red when they take damage, and go back to their original
color after their invincibility period.