Added game flow manager and Pausing
This is when I've begun to implement the game manager. I haven't fully
implemented it yet as the meta systems of the game aren't yet
implemented, but this is the infrastructure needed to implement these
systems. I have however implemented pausing. If you were wondering why
the player's shooter is disabled on pause, it's because otherwise the
player can shoot an extra shot they shouldn't be able to. I've also
turned the enemies from the last commit into prefabs.