Added score system and UI + automatic waves
This is a big one - the first thing I did was make waves start spawning
automatically once the game begins. This is a temporary measure as in
the future waves will be spawned only once you begin the game through
the main menu. I then added a score value to the enemy controller
component which I then used to implement the score manager. I then
added text on screen which shows the current health, score, and wave
number. Finally I implemented the high score file which the above
detailed numbers are written to when the game ends.